backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.0.0
pydantic-settings==2.5.2